Update main.scala
I change the 'eval' method because the Int return type was generating inconsistencies, so I make it Double.

+++ Project 2/main.scala
@@ -28,10 +28,10 @@ object Methods
   }
 
 
-  def eval(root: Tree[Char]): Int =
+  def eval(root: Tree[Char]): Double =
   {
 
-    def go(node: Tree[Char]): Int = node match
+    def go(node: Tree[Char]): Double = node match
     {
 
       case Branch(value, left, right) =>
